Added rename_file and delete_file.
